Weblog-By Django/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
defusedxml==0.7.1
Django==4.0
django-ckeditor==6.2.0
django-comments-dab==2.7.1
django-crispy-forms==1.13.0
django-gravatar2==1.4.4
django-js-asset==1.2.2
django-utils-six==2.0
django-widget-tweaks==1.4.9
idna==3.3
oauthlib==3.1.1
pycparser==2.21
PyJWT==2.3.0
python-decouple==3.5
python3-openid==3.2.0
requests==2.26.0
requests-oauthlib==1.3.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
urllib3==1.26.7
